refactor v2 eval keyed collection ops
